Initial commit: most of a semi-working disassembler.
I had *all* of a semi-working disassembler and then I accidentally
deleted everything.  Thankfully I still had ARM.hs open in vim, so I
was able to save it again, but the part that does the actual
disassembling is gone.  Fuck.

Note to self: stick stuff in version control from the start.