Make a few minor stylistic changes.
Consolidating the two additions in getPC was a bigger decision than it
should have been.  To me, it should be "baseAddress + bytesRead + 4",
which is kind of why I had it the old way, or at least "bytesRead +
baseAddress + 4", but "(+ baseAddress + 4)" doesn't work.  Having the 4
in the middle is fucking weird, but I guess splitting the addition up
is weirder.

The newFuture line was eighty characters long and I go with a limit of
seventy-nine.

Disassembling a binary, to me, is one conceptual chunk and should be
the entirety of a function.  Disassembling *and* printing is two
things, and it's the main function that wants to do both.