Actually disassemble Thumb, sort of, and only for Pokémon FireRed.
A lot of the sections it's producing overlap.  It's not supposed to
branch addresses as their own sections if they fall in the middle of
other sections!  However, it also assumes that you're not going to
branch into the middle of a section from outside of that section,
which I assume is where it's going wrong.  Dammit.

Also.  Yes.  There's no way to specify disassembly points and it
still can't follow BX so I manually special-cased firered.gba for
the time being so I can push and go to bed.